reorganization and bug fixes
