+++ scripts/0_init_env.R
@@ -5,7 +5,7 @@
 ## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 # Vector of required packages
-package_list <- c("tidyverse", "here", "lubridate")
+package_list <- c("tidyverse", "here", "lubridate", "dplyr")
 
 # Check if package is installed and load
 for (package in package_list) {
